lightssl fix MOD

+++ src/lightdefs.h
@@ -61,5 +61,6 @@ extern const u64 SHA_K[80];
 #define SHA_AddLength(c, l) (SHA_T(c, l) && SHA_L(c))
 
 // Imitate pythons %. -1 % 5 = 4, not -1
-#define MOD(n, m) ((n % m) + m) % m
+#define MOD(n, m) (((int)n % (int)m) + (int)m) % (int)m
+
 #endif
